# shared/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ada Fitness Tracker API
  description: API for the Ada fitness and nutrition tracking application
  version: 1.0.0
  contact:
    email: contact@ada-fitness.app
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ada-fitness.app
    description: Production server
...
paths:
...
  /nutrition/meals:
    get:
      tags: [Nutrition]
      summary: List meals
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MealSummary'
    post:
      tags: [Nutrition]
      summary: Create meal
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MealCreate'
      responses:
        '201':
          description: Created meal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meal'
  /nutrition/foods/barcode/{barcode}:
    get:
      tags: [Nutrition]
      summary: Lookup product by barcode
      security:
        - bearerAuth: []
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenFoodFactsProduct'
        '404':
          description: Product not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Meal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        meal_type:
          type: string
        timestamp:
          type: string
          format: date-time
        total_calories:
          type: integer
        total_protein_g:
          type: number
        total_carbs_g:
          type: number
        total_fat_g:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/FoodItem'
    MealSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        meal_type:
          type: string
        timestamp:
          type: string
          format: date-time
        total_calories:
          type: integer
        items_count:
          type: integer
    MealCreate:
      type: object
      required: [name, timestamp, items]
      properties:
        name:
          type: string
        meal_type:
          type: string
          default: other
        timestamp:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/FoodItemCreate'
    FoodItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        source:
          type: string
        grams:
          type: number
        calories:
          type: integer
        protein_g:
          type: number
        carbs_g:
          type: number
        fat_g:
          type: number
    FoodItemCreate:
      type: object
      required: [name, grams, calories, protein_g, carbs_g, fat_g]
      properties:
        name:
          type: string
        source:
          type: string
          default: manual
        grams:
          type: number
        calories:
          type: integer
        protein_g:
          type: number
        carbs_g:
          type: number
        fat_g:
          type: number
    OpenFoodFactsProduct:
      type: object
      properties:
        barcode:
          type: string
        name:
          type: string
        brand:
          type: string
        calories_per_100g:
          type: number
        protein_per_100g:
          type: number
        carbs_per_100g:
          type: number
        fat_per_100g:
          type: number
...